src/embeddings-server/requirements.txt
fastapi==0.114.1
uvicorn==0.30.6
chromadb==0.5.7
python-dotenv==1.0.1
sentence-transformers==3.1.1
httpx==0.27.2
llama-index-core==0.11.16
llama-index-readers-file==0.2.2